Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 int main(){
-    float a;
+    int a;
     scanf("%d",&a);
 
-    float b;
+    int b;
     scanf("%d",&b);
 
-    float c;
+    int c;
     scanf("%d",&c);
 
     float average = (a+b+c)/3;
